requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.5
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
Flask==2.0.1
Flask-Caching==1.10.1
hexbytes==0.2.2
idna==3.2
ipfshttpclient==0.8.0a2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
lru-dict==1.1.7
MarkupSafe==2.0.1
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
numpy==1.21.2
pandas==1.3.2
parsimonious==0.8.1
protobuf==3.17.3
pycryptodome==3.10.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rlp==2.0.1
six==1.16.0
toolz==0.11.1
typing-extensions==3.10.0.2
url-normalize==1.4.3
urllib3==1.26.6
varint==1.0.2
web3==5.23.1
websockets==9.1
Werkzeug==2.0.1
yarl==1.6.3
